+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Implemented file locking to avoid data races (guaranteed for Linux/MacOSX)
 - Implemented ```pull``` command to fetch and save multiple thing models at once
 - Implemented setting CORS options for API
+- Implemented fetching a TM by a \[partial\] semantic version also in REST API
 - Print information about used config file in `help`
 - Implemented a 'tmc' remote type, which uses our own REST API as the underlying TM storage
 - Added 'filter.name' parameter to REST API '/inventory' listing
@@ -17,6 +18,7 @@
 - ```list```: allows now listing by name pattern
 - ```serve```: separate configuration of the remote(s) to be served from the target remote for push
 - ```fetch```: ```--output``` now accepts only a target folder to save TM to, ```--with-path``` has been removed
+- enable/disable logging is now done only by setting a loglevel
 - `list` and `pull`: match given name pattern as a prefix by complete path parts
 - `list`: changed output format: put NAME column first, renamed PATH column to MPN
 - `versions`: changed output format: renamed PATH column to ID
@@ -24,6 +26,7 @@
 ### Fixed
 
 - count only enabled remotes when checking if empty remote specification is unambiguous
+- make fetch by partial semantic version match the most recent version beginning with given string
 - (BREAKING!) ```push```: file hash calculation has been made more reliable and idempotent. Consequently, some files if pushed to TMC, may receive a new version hash, despite no change in contents
 - ```fetch```: fixed "Unable to parse TMID..." error when fetching an official TM by content hash
 - prevent ```serve``` from using one of remotes from config as push target when '-r' or '-d' are given